[REFACTOR] Update tests to match actual Glimmer component API
The current test setup uses a much older version of what Glimmer
components ostensibly looked like. The EmberishGlimmerComponent in them
has a ton of lifecycle hooks, does not assign args to `args`, has a
`recompute` function, etc. It also uses an internal component manager
rather than a public facing one, unlike Glimmer components.

This PR refactors tests to use the custom manager setup and to make them
work effectively like Glimmer components do in practice.
