+++ packages/@glimmer/integration-tests/lib/suites/emberish-components.ts
@@ -1,161 +1,13 @@
 import { RenderTest, Count } from '../render-test';
 import { test } from '../test-decorator';
 import { SimpleElement } from '@simple-dom/interface';
-import { EmberishGlimmerComponent, EmberishCurlyComponent } from '../components';
-import { strip } from '../test-helpers/strings';
-import { assertElementShape, classes, assertEmberishElement } from '../dom/assertions';
+import { EmberishCurlyComponent } from '../components';
+import { classes, assertEmberishElement } from '../dom/assertions';
 import { assertElement, toInnerHTML } from '../dom/simple-utils';
-import { EmberishGlimmerArgs } from '../components/emberish-glimmer';
 import { equalTokens } from '../snapshot';
 
 export class EmberishComponentTests extends RenderTest {
   static suiteName = 'Emberish';
-
-  @test({ kind: 'glimmer' })
-  '[BUG: #644 popping args should be balanced]'() {
-    class MainComponent extends EmberishGlimmerComponent {
-      salutation = 'Glimmer';
-    }
-    this.registerComponent(
-      'Glimmer',
-      'Main',
-      '<div><HelloWorld @name={{salutation}} /></div>',
-      MainComponent
-    );
-    this.registerComponent('Glimmer', 'HelloWorld', '<h1>Hello {{@name}}!</h1>');
-    this.render('<Main />');
-    this.assertHTML('<div><h1>Hello Glimmer!</h1></div>');
-  }
-
-  @test({ kind: 'glimmer' })
-  'Only one arg reference is created per argument'() {
-    let count = 0;
-
-    this.registerHelper('count', () => count++);
-
-    class MainComponent extends EmberishGlimmerComponent {
-      salutation = 'Glimmer';
-    }
-    this.registerComponent(
-      'Glimmer',
-      'Main',
-      '<div><Child @value={{count}} /></div>',
-      MainComponent
-    );
-    this.registerComponent('Glimmer', 'Child', '{{@value}} {{this.value}}');
-    this.render('<Main />');
-    this.assertHTML('<div>0 0</div>');
-  }
-
-  @test({ kind: 'glimmer' })
-  '[BUG] Gracefully handles application of curried args when invoke starts with 0 args'() {
-    class MainComponent extends EmberishGlimmerComponent {
-      salutation = 'Glimmer';
-    }
-    this.registerComponent(
-      'Glimmer',
-      'Main',
-      '<div><HelloWorld @a={{@a}} as |wat|>{{wat}}</HelloWorld></div>',
-      MainComponent
-    );
-    this.registerComponent('Glimmer', 'HelloWorld', '{{yield (component "A" a=@a)}}');
-    this.registerComponent('Glimmer', 'A', 'A {{@a}}');
-    this.render('<Main @a={{a}} />', { a: 'a' });
-    this.assertHTML('<div>A a</div>');
-    this.assertStableRerender();
-    this.rerender({ a: 'A' });
-    this.assertHTML('<div>A A</div>');
-    this.assertStableNodes();
-  }
-
-  @test({ kind: 'glimmer' })
-  'Static block component helper'() {
-    this.registerComponent(
-      'Glimmer',
-      'A',
-      'A {{#component "B" arg1=@one arg2=@two arg3=@three}}{{/component}}'
-    );
-    this.registerComponent('Glimmer', 'B', 'B {{@arg1}} {{@arg2}} {{@arg3}}');
-    this.render('<A @one={{first}} @two={{second}} @three={{third}} />', {
-      first: 1,
-      second: 2,
-      third: 3,
-    });
-    this.assertHTML('A B 1 2 3');
-    this.assertStableRerender();
-    this.rerender({ first: 2, second: 3, third: 4 });
-    this.assertHTML('A B 2 3 4');
-    this.assertStableNodes();
-  }
-
-  @test({ kind: 'glimmer' })
-  'Static inline component helper'() {
-    this.registerComponent('Glimmer', 'A', 'A {{component "B" arg1=@one arg2=@two arg3=@three}}');
-    this.registerComponent('Glimmer', 'B', 'B {{@arg1}} {{@arg2}} {{@arg3}}');
-    this.render('<A @one={{first}} @two={{second}} @three={{third}} />', {
-      first: 1,
-      second: 2,
-      third: 3,
-    });
-    this.assertHTML('A B 1 2 3');
-    this.assertStableRerender();
-    this.rerender({ first: 2, second: 3, third: 4 });
-    this.assertHTML('A B 2 3 4');
-    this.assertStableNodes();
-  }
-
-  @test({ kind: 'glimmer' })
-  'top level in-element'() {
-    this.registerComponent('Glimmer', 'Foo', '<Bar data-bar={{@childName}} @data={{@data}} />');
-    this.registerComponent('Glimmer', 'Bar', '<div ...attributes>Hello World</div>');
-
-    let el = this.delegate.getInitialElement();
-
-    this.render(
-      strip`
-    {{#each components key="id" as |c|}}
-      {{#in-element c.mount}}
-        {{component c.name childName=c.child data=c.data}}
-      {{/in-element}}
-    {{/each}}
-    `,
-      { components: [{ name: 'Foo', child: 'Bar', mount: el, data: { wat: 'Wat' } }] }
-    );
-
-    let first = assertElement(el.firstChild);
-
-    assertElementShape(first, 'div', { 'data-bar': 'Bar' }, 'Hello World');
-    this.rerender({ components: [{ name: 'Foo', child: 'Bar', mount: el, data: { wat: 'Wat' } }] });
-    assertElementShape(first, 'div', { 'data-bar': 'Bar' }, 'Hello World');
-  }
-
-  @test({ kind: 'glimmer' })
-  'recursive component invocation'() {
-    let counter = 0;
-
-    class RecursiveInvoker extends EmberishGlimmerComponent {
-      id: number;
-
-      get showChildren() {
-        return this.id < 3;
-      }
-
-      constructor(args: EmberishGlimmerArgs) {
-        super(args);
-        this.id = ++counter;
-      }
-    }
-
-    this.registerComponent(
-      'Glimmer',
-      'RecursiveInvoker',
-      '{{id}}{{#if showChildren}}<RecursiveInvoker />{{/if}}',
-      RecursiveInvoker
-    );
-
-    this.render('<RecursiveInvoker />');
-    this.assertHTML('123<!---->');
-  }
 
   @test
   'Element modifier with hooks'(assert: Assert, count: Count) {
